PlanejNecessidades

+++ XlToDb/Model/CuboEstoque.cs
@@ -20,6 +20,6 @@ namespace XlToDb.Model
 
         [Display(Name = "Quantidade")]
         [DisplayFormat(DataFormatString = "{0:N0}")]
-        public int Quantidade { get; set; }
+        public float Quantidade { get; set; }
     }
 }
